InfoArea test First test for the infoarea. menu and widget separated into two different files to avoid circular import. Works but need to be redefined: widget need a ScreenArea (or child) as reference to work with coordinates.
